workers/polymarket/src/pyproject.toml
[project]
name = "polymarket-market-maker-bot"
version = "0.1.0"
description = "Professional Polymarket CLOB market-making bot with inventory management, spread farming, and optimized cancel/replace cycles."
authors = [{ name = "Polymarket Market Maker" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=12.0",
    "httpx>=0.27.0",
    "web3>=6.15.0",
    "eth-account>=0.10.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.3.0",
    "ruff>=0.6.8",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
target-version = "py311"

